added remaining info fields in agency profile and the code to display them, icons are still needed tho

+++ app/src/main/java/com/example/myapplication/Model/Agency.java
@@ -27,7 +27,7 @@ public class Agency extends User implements Serializable {
                 userType);
     }
 
-    public Agency(int patentNumber,
+    public Agency(long patentNumber,
             String managerFullName, String agencyName) {
         super();
     }
